Initial Commit
Uploaded a Java program to find a minimum square to contain K number of x,y coordinates.  Created for a programming test.
Lawn.js is a lawn game created for a game jam, associated html and css files for testing purposes. 
Simple_Neural_Network.ipynb was created as a result of reading a how to on neural networks. More work on these will come in the future.